clean websocket
